Notice refactor -- Move NoticeExporter interface to domain
