+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/CannotConstructDataProviderNotice.java
@@ -31,8 +31,7 @@ public class CannotConstructDataProviderNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
 }
