+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/CannotDownloadArchiveFromNetworkNotice.java
@@ -40,8 +40,7 @@ public class CannotDownloadArchiveFromNetworkNotice extends ErrorNotice {
 
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
 }
