+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/FloatFieldValueOutOfRangeNotice.java
@@ -46,8 +46,7 @@ public class FloatFieldValueOutOfRangeNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
