+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/IntegerFieldValueOutOfRangeNotice.java
@@ -46,8 +46,7 @@ public class IntegerFieldValueOutOfRangeNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
