+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/InvalidRowLengthNotice.java
@@ -38,8 +38,7 @@ public class InvalidRowLengthNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
