+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/InvalidTimeNotice.java
@@ -43,7 +43,7 @@ public class InvalidTimeNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter) throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
 }
