+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/InvalidTimezoneNotice.java
@@ -36,8 +36,7 @@ public class InvalidTimezoneNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
