+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/MissingRequiredFileNotice.java
@@ -30,8 +30,7 @@ public class MissingRequiredFileNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
 }
