+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/MissingRequiredValueNotice.java
@@ -34,8 +34,7 @@ public class MissingRequiredValueNotice extends ErrorNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
-            throws IOException {
+    public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
