+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/warning/NonAsciiOrNonPrintableCharNotice.java
@@ -35,7 +35,7 @@ public class NonAsciiOrNonPrintableCharNotice extends WarningNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
+    public void export(final NoticeExporter exporter)
             throws IOException {
         exporter.export(this);
     }
