+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/warning/NonStandardHeaderNotice.java
@@ -38,7 +38,7 @@ public class NonStandardHeaderNotice extends WarningNotice {
     }
 
     @Override
-    public void export(ValidationResultRepository.NoticeExporter exporter)
+    public void export(final NoticeExporter exporter)
             throws IOException {
         exporter.export(this);
     }
