+++ usecase/src/main/java/org/mobilitydata/gtfsvalidator/usecase/port/ValidationResultRepository.java
@@ -38,54 +38,4 @@ public interface ValidationResultRepository {
     Collection<Notice> getAll();
 
     NoticeExporter getExporter(boolean outputAsProto, String outputPath) throws IOException;
-
-    // an interface that will be implemented in different flavors: proto, json, string, ...
-    interface NoticeExporter {
-
-        String getExtension();
-
-        void exportBegin() throws IOException;
-
-        void exportEnd() throws IOException;
-
-        void export(NonStandardHeaderNotice toExport) throws IOException;
-
-        void export(InputZipContainsFolderNotice toExport) throws IOException;
-
-        void export(NonAsciiOrNonPrintableCharNotice toExport) throws IOException;
-
-        void export(CannotConstructDataProviderNotice toExport) throws IOException;
-
-        void export(CannotDownloadArchiveFromNetworkNotice toExport) throws IOException;
-
-        void export(CannotParseFloatNotice toExport) throws IOException;
-
-        void export(CannotParseIntegerNotice toExport) throws IOException;
-
-        void export(CannotUnzipInputArchiveNotice toExport) throws IOException;
-
-        void export(FloatFieldValueOutOfRangeNotice toExport) throws IOException;
-
-        void export(IntegerFieldValueOutOfRangeNotice toExport) throws IOException;
-
-        void export(InvalidRowLengthNotice toExport) throws IOException;
-
-        void export(InvalidTimezoneNotice toExport) throws IOException;
-
-        void export(InvalidUrlNotice toExport) throws IOException;
-
-        void export(MissingHeaderNotice toExport) throws IOException;
-
-        void export(MissingRequiredFileNotice toExport) throws IOException;
-
-        void export(MissingRequiredValueNotice toExport) throws IOException;
-
-        void export(InvalidColorNotice toExport) throws IOException;
-
-        void export(ExtraFileFoundNotice toExport) throws IOException;
-
-        void export(InvalidTimeNotice toExport) throws IOException;
-
-        void export(CannotParseDateNotice toExport) throws IOException;
-    }
 }
